Refactor epic + banner components to share logic and models (#127)
* refactor epic + banner components to share logic and models

* make TestsList a functional component
