+++ public/src/components/channelManagement/epicTests/epicTestsForm.tsx
@@ -1,27 +1,19 @@
 import React from 'react';
-import update from 'immutability-helper';
 import {createStyles, Theme, withStyles, WithStyles, Typography} from "@material-ui/core";
 import EpicTestEditor from './epicTestEditor';
-import TestActionBar from '../testActionBar';
-import CircularProgress from '@material-ui/core/CircularProgress';
 import {Region} from "../../../utils/models";
-import {
-  fetchFrontendSettings,
-  FrontendSettingsType,
-  saveFrontendSettings,
-  requestLock,
-  requestTakeControl,
-  requestUnlock,
-  archiveEpicTest
-} from "../../../utils/requests";
-import EpicTestsList from "./epicTestsList";
+
+import TestsList from "../testsList";
 import {
   UserCohort,
   Cta,
   ArticlesViewedSettings,
-  LockStatus,
-  ModifiedTests,
+  Test,
 } from "../helpers/shared";
+import {InnerComponentProps, updateTest} from "../testEditor";
+import TestsForm from "../testEditor";
+import {FrontendSettingsType} from "../../../utils/requests";
+import {MaxEpicViewsDefaults} from "./maxEpicViewsEditor";
 
 export enum TickerEndType {
   unlimited = 'unlimited',
@@ -61,7 +53,7 @@ export interface MaxEpicViews {
   minDaysBetweenViews: number
 }
 
-export interface EpicTest {
+export interface EpicTest extends Test {
   name: string,
   nickname?: string,
   isOn: boolean,
@@ -81,25 +73,24 @@ export interface EpicTest {
   articlesViewedSettings?: ArticlesViewedSettings
 }
 
-interface EpicTests {
-  tests: EpicTest[]
-}
-
-interface DataFromServer {
-  value: EpicTests,
-  version: string,
-  status: LockStatus,
-  userEmail: string,
-};
-
-type EpicTestsFormState = EpicTests & {
-  version: string | null,
-  selectedTestName?: string,
-  editMode: boolean,
-  lockStatus: LockStatus,
-  modifiedTests: ModifiedTests,
-  timeoutAlertId: number | null,  // A timeout for warning about being open for edit for too long
-};
+const createDefaultEpicTest = (newTestName: string, newTestNickname: string): EpicTest => ({
+  name: newTestName,
+  nickname: newTestNickname,
+  isOn: false,
+  locations: [],
+  tagIds: [],
+  sections: [],
+  excludedTagIds: [],
+  excludedSections: [],
+  alwaysAsk: false,
+  maxViews: MaxEpicViewsDefaults,
+  userCohort: UserCohort.AllNonSupporters,  // matches the default in dotcom
+  isLiveBlog: false,
+  hasCountryName: false,
+  variants: [],
+  highPriority: false, // has been removed from form, but might be used in future
+  useLocalViewLog: false,
+});
 
 const styles = ({ spacing, typography }: Theme) => createStyles({
   testListAndEditor: {
@@ -122,270 +113,64 @@ const styles = ({ spacing, typography }: Theme) => createStyles({
   }
 });
 
-interface EpicTestFormProps extends WithStyles<typeof styles> {}
+interface Props extends InnerComponentProps<EpicTest>, WithStyles<typeof styles> {}
 
-class EpicTestsForm extends React.Component<EpicTestFormProps, EpicTestsFormState> {
-  state: EpicTestsFormState;
-
-  constructor(props: EpicTestFormProps) {
-    super(props);
-    this.state = {
-      tests: [],
-      version: null,
-      selectedTestName: undefined,
-      editMode: false,
-      lockStatus: { locked: false },
-      modifiedTests: {},
-      timeoutAlertId: null,
-    };
-  };
-
-  componentWillMount(): void {
-    this.fetchStateFromServer()
-  };
-
-  fetchStateFromServer = (): void => {
-    fetchFrontendSettings(FrontendSettingsType.epicTests)
-      .then((serverData: DataFromServer) => {
-        const editMode = serverData.status.email === serverData.userEmail;
-
-        this.updateWarningTimeout(editMode);
-
-        this.setState({
-          ...serverData.value,
-          version: serverData.version,
-          lockStatus: serverData.status,
-          editMode: editMode,
-          modifiedTests: {}
-        });
-      });
-  };
-
-  // Maintains an alert if tool is left open for edit for 20 minutes
-  updateWarningTimeout = (editMode: boolean): void => {
-    if (editMode) {
-      if (this.state.timeoutAlertId) {
-        window.clearTimeout(this.state.timeoutAlertId);
-      }
-
-      const timeoutAlertId = window.setTimeout(() => {
-        alert("You've had this editing session open for 20 minutes - if you leave it much longer then you may lose any unsaved work!\nIf you've finished then please either save or cancel.");
-        this.setState({ timeoutAlertId: null });
-      }, 60 * 20 * 1000);
-
-      this.setState({ timeoutAlertId });
-
-    } else if (this.state.timeoutAlertId) {
-      window.clearTimeout(this.state.timeoutAlertId);
-      this.setState({ timeoutAlertId: null });
-    }
-  };
-
-  cancel = (): void => {
-    requestUnlock(FrontendSettingsType.epicTests).then( response =>
-      response.ok ? this.fetchStateFromServer() : alert("Error - can't request lock!")
-    )
-  };
-
-  save = (): void => {
-    const testsToArchive: EpicTest[] = this.state.tests.filter(test =>
-      this.state.modifiedTests[test.name] && this.state.modifiedTests[test.name].isArchived
-    );
-
-    Promise.all(testsToArchive.map(test => archiveEpicTest(test))).then(results => {
-      const notOk = results.some(result => !result.ok);
-      const numTestsToArchive = testsToArchive.length;
-      if (notOk) {
-        alert(`Failed to archive ${numTestsToArchive} test${numTestsToArchive !== 1 ? 's' : ''}`);
-      } else {
-        const updatedTests: EpicTest[] = this.state.tests.filter(test => {
-          const modifiedTestData = this.state.modifiedTests[test.name];
-          return !(modifiedTestData && (modifiedTestData.isDeleted || modifiedTestData.isArchived));
-        });
-
-        const postData = {
-          version: this.state.version,
-          value: {
-            tests: updatedTests
-          }
-        };
-
-        saveFrontendSettings(FrontendSettingsType.epicTests, postData)
-          .then(resp => {
-            if (!resp.ok) {
-              resp.text().then(msg => alert(msg));
-            }
-            this.fetchStateFromServer();
-          })
-          .catch((resp) => {
-            alert('Error while saving');
-            this.fetchStateFromServer();
-          });
-      }
-    });
-  };
-
-  onTestsChange = (updatedTests: EpicTest[], modifiedTestName?: string): void => {
-    if (modifiedTestName && !this.state.modifiedTests[modifiedTestName]) {
-      this.setState({
-        modifiedTests: {
-          ...this.state.modifiedTests,
-          [modifiedTestName]: {
-            isValid: true, // not already modified, assume it's valid until told otherwise
-            isDeleted: false,
-            isArchived: false,
-            isNew: !this.state.tests.some(test => test.name === modifiedTestName)
-          }
-        }
-      })
-    }
-
-    this.setState({
-      tests: updatedTests
-    });
-  };
-
-  onTestChange = (updatedTest: EpicTest): void => {
-    const updatedTests = this.state.tests.map(test => test.name === updatedTest.name ? updatedTest : test);
-    this.onTestsChange(updatedTests, updatedTest.name);
-  };
-
-  onTestErrorStatusChange = (testName: string) => (isValid: boolean): void => {
-    if (this.state.modifiedTests[testName]) {
-      this.setState({
-        modifiedTests: {
-          ...this.state.modifiedTests,
-          [testName]: {
-            ...this.state.modifiedTests[testName],
-            isValid
-          }
-        }
-      });
-    }
-  };
-
-  onTestDelete = (testName: string): void => {
-    const updatedState = this.state.modifiedTests[testName] ?
-      { ...this.state.modifiedTests[testName], isDeleted: true } :
-      {
-        isValid: true,
-        isDeleted: true,
-        isNew: false,
-        isArchived: false
-      };
-
-    this.setState({
-      modifiedTests: {
-        ...this.state.modifiedTests,
-        [testName]: updatedState
-      }
-    });
-  };
-
-  onTestArchive = (testName: string): void => {
-    const updatedState = this.state.modifiedTests[testName] ?
-    { ...this.state.modifiedTests[testName], isArchived: true } :
-    {
-      isValid: true,
-      isDeleted: false,
-      isNew: false,
-      isArchived: true
-    };
-
-    this.setState({
-      modifiedTests: {
-        ...this.state.modifiedTests,
-        [testName]: updatedState
-      }
-    });
-  }
-
-  onSelectedTestName = (testName: string): void => {
-      this.setState({
-        selectedTestName: testName
-      })
-  };
-
-  requestEpicTestsLock = () => {
-    requestLock(FrontendSettingsType.epicTests).then(response =>
-      response.ok ? this.fetchStateFromServer() : alert("Error - can't request lock!")
-    );
-  };
-
-  requestTakeControl = () => {
-    requestTakeControl(FrontendSettingsType.epicTests).then(response =>
-      response.ok ? this.fetchStateFromServer() : alert("Error - can't take back control!")
-    );
-  };
-
+class EpicTestsForm extends React.Component<Props> {
   render(): React.ReactNode {
     const { classes } = this.props;
     const listAndEditorClassNames = [
       classes.testListAndEditor,
-      this.state.editMode && classes.editModeBorder,
-      !this.state.editMode && classes.readOnlyModeBorder
+      this.props.editMode && classes.editModeBorder,
+      !this.props.editMode && classes.readOnlyModeBorder
     ].join(' ');
 
     return (
-      <>
-        <div>
-          {
-            this.state.version ?
-              <>
-                <TestActionBar
-                  modifiedTests={this.state.modifiedTests}
-                  lockStatus={this.state.lockStatus}
-                  editMode={this.state.editMode}
-                  requestTakeControl={this.requestTakeControl}
-                  requestLock={this.requestEpicTestsLock}
-                  cancel={this.cancel}
-                  save={this.save}
-                />
-                <div className={listAndEditorClassNames}>
-                  <EpicTestsList
-                    tests={this.state.tests}
-                    modifiedTests={this.state.modifiedTests}
-                    selectedTestName={this.state.selectedTestName}
-                    onUpdate={this.onTestsChange}
-                    onSelectedTestName={this.onSelectedTestName}
-                    editMode={this.state.editMode}
-                  />
+      <div className={listAndEditorClassNames}>
+        <TestsList<EpicTest>
+          tests={this.props.tests}
+          modifiedTests={this.props.modifiedTests}
+          selectedTestName={this.props.selectedTestName}
+          onUpdate={this.props.onTestsChange}
+          createDefaultTest={createDefaultEpicTest}
+          onSelectedTestName={this.props.onSelectedTestName}
+          editMode={this.props.editMode}
+        />
 
-                  {this.state.selectedTestName ? this.state.tests.map(test =>
-                    (<EpicTestEditor
-                      test={this.state.tests.find(test => test.name === this.state.selectedTestName)}
-                      hasChanged={!!this.state.modifiedTests[test.name]}
-                      onChange={this.onTestChange}
-                      onValidationChange={this.onTestErrorStatusChange(test.name)}
-                      visible={test.name === this.state.selectedTestName}
-                      key={test.name}
-                      editMode={this.state.editMode}
-                      onDelete={this.onTestDelete}
-                      onArchive={this.onTestArchive}
-                      isDeleted={this.state.modifiedTests[test.name] && this.state.modifiedTests[test.name].isDeleted}
-                      isArchived={this.state.modifiedTests[test.name] && this.state.modifiedTests[test.name].isArchived}
-                      isNew={this.state.modifiedTests[test.name] && this.state.modifiedTests[test.name].isNew}
-                      createTest={(newTest: EpicTest) => {
-                        const newTests = [...this.state.tests, newTest];
-                        this.onTestsChange(newTests, newTest.name)
-                      }}
-                      testNames={this.state.tests.map(test => test.name)}
-                      testNicknames={
-                        this.state.tests
-                          .map(test => test.nickname)
-                          .filter(nickname => !!nickname) as string[]
-                      }
-                    />)
-                  ) : (<Typography className={classes.viewText}>Click on a test on the left to view contents.</Typography>)}
-                </div>
-              </>
-               :
-              <CircularProgress/>
-          }
-        </div>
-      </>
+        {this.props.selectedTestName ? this.props.tests.map(test =>
+          (<EpicTestEditor
+            test={this.props.tests.find(test => test.name === this.props.selectedTestName)}
+            hasChanged={!!this.props.modifiedTests[test.name]}
+            onChange={updatedTest =>
+              this.props.onTestsChange(
+                updateTest(this.props.tests, updatedTest),
+                updatedTest.name
+              )
+            }
+            onValidationChange={this.props.onTestErrorStatusChange(test.name)}
+            visible={test.name === this.props.selectedTestName}
+            key={test.name}
+            editMode={this.props.editMode}
+            onDelete={this.props.onTestDelete}
+            onArchive={this.props.onTestArchive}
+            isDeleted={this.props.modifiedTests[test.name] && this.props.modifiedTests[test.name].isDeleted}
+            isArchived={this.props.modifiedTests[test.name] && this.props.modifiedTests[test.name].isArchived}
+            isNew={this.props.modifiedTests[test.name] && this.props.modifiedTests[test.name].isNew}
+            createTest={(newTest: EpicTest) => {
+              const newTests = [...this.props.tests, newTest];
+              this.props.onTestsChange(newTests, newTest.name)
+            }}
+            testNames={this.props.tests.map(test => test.name)}
+            testNicknames={
+              this.props.tests
+                .map(test => test.nickname)
+                .filter(nickname => !!nickname) as string[]
+            }
+          />)
+        ) : (<Typography className={classes.viewText}>Click on a test on the left to view contents.</Typography>)}
+      </div>
     )
   }
 }
 
-export default withStyles(styles)(EpicTestsForm);
+const styled = withStyles(styles)(EpicTestsForm);
+export default TestsForm(styled, FrontendSettingsType.epicTests)
