+++ public/src/components/channelManagement/tickerEditor.tsx
@@ -11,7 +11,7 @@ import {
 } from "@material-ui/core";
 import Switch from "@material-ui/core/Switch";
 import FormControlLabel from "@material-ui/core/FormControlLabel";
-import {TickerCountType, TickerEndType, TickerSettings} from "./epicTests/epicTestsForm";
+import {TickerCountType, TickerEndType, TickerSettings} from "./helpers/shared";
 import EditableTextField from "./helpers/editableTextField"
 import {onFieldValidationChange, ValidationStatus} from "./helpers/validation";
