+++ public/src/main.tsx
@@ -6,6 +6,7 @@ import Switchboard from './components/switchboard';
 import ContributionTypesForm from './components/contributionTypes';
 import AmountsForm from './components/amounts/amounts';
 import EpicTestsForm from './components/channelManagement/epicTests/epicTestsForm';
+import BannerTestsForm from "./components/channelManagement/bannerTests/bannerTestsForm";
 import {createStyles, Theme, WithStyles, withStyles} from '@material-ui/core/styles';
 import CssBaseline from '@material-ui/core/CssBaseline';
 import AppBar from '@material-ui/core/AppBar';
@@ -14,7 +15,6 @@ import NavDrawer from "./components/drawer";
 import Toolbar from "@material-ui/core/Toolbar";
 import Typography from "@material-ui/core/Typography";
 import IndexPage from "./components/indexPage";
-import BannerTestsForm from "./components/channelManagement/bannerTests/bannerTestsForm";
 
 const styles = ({ palette, spacing, mixins, typography, transitions }: Theme) => createStyles({
   root: {
